+++ packages/std/src/math/decimal.rs
@@ -95,15 +95,15 @@ impl Decimal {
     }
 }
 
-impl Fraction<u128> for Decimal {
+impl Fraction<Uint128> for Decimal {
     #[inline]
-    fn numerator(&self) -> u128 {
-        self.0.u128()
+    fn numerator(&self) -> Uint128 {
+        self.0
     }
 
     #[inline]
-    fn denominator(&self) -> u128 {
-        Self::DECIMAL_FRACTIONAL.u128()
+    fn denominator(&self) -> Uint128 {
+        Self::DECIMAL_FRACTIONAL
     }
 
     /// Returns the multiplicative inverse `1/d` for decimal `d`.
@@ -360,8 +360,14 @@ mod tests {
     #[test]
     fn decimal_implements_fraction() {
         let fraction = Decimal::from_str("1234.567").unwrap();
-        assert_eq!(fraction.numerator(), 1_234_567_000_000_000_000_000u128);
-        assert_eq!(fraction.denominator(), 1_000_000_000_000_000_000u128);
+        assert_eq!(
+            fraction.numerator(),
+            Uint128::from(1_234_567_000_000_000_000_000u128)
+        );
+        assert_eq!(
+            fraction.denominator(),
+            Uint128::from(1_000_000_000_000_000_000u128)
+        );
     }
 
     #[test]